requirements.txt
fastapi
uvicorn
openai
supabase
python-dotenv
requests
pandas
